API

This package implements all of the required JSMD Interfaces for Ephemeris files, which are already well-documented in JSMDInterfaces.

CalcephEphemeris.CalcephProvider — Type

CalcephProvider(file::String)
CalcephProvider(files::Vector{String})

Create a CalcephProvider instance by loading a single or multiples ephemeris kernel files specified by files.

Note

Once the object is created kernels cannot be added nor removed from the generated CalcephProvider instance.

Example

julia> eph1 = CalcephProvider("PATH_TO_KERNEL")
1-kernel CalcephProvider
 "PATH_TO_KERNEL"

julia> eph2 = CalcephProvider(["PATH_TO_KERNEL_1", "PATH_TO_KERNEL_2"])
2-kernel CalcephProvider:
 "PATH_TO_KERNEL_1"
 "PATH_TO_KERNEL_2"